Driver Station Read Controls
Read two encoders from the Driver Station using the TI MSP-430
LaunchPad. Information is passed as 16 "buttons" which are split into 2
x 8-bit unsigned integers. Additional Code to process this 16 bit
"button" array to convert it to an integer. Firmware is set so that
encoders never pass a value less than 0 or more than 180.